refactor: main.ts
