+++ src/app.module.ts
@@ -5,6 +5,7 @@ import {
   RequestMethod,
 } from '@nestjs/common';
 import { ApiModule } from '@src/apis/api.module';
+import { AppService } from '@src/app.service';
 import { CoreModule } from '@src/core/core.module';
 import { HttpExceptionModule } from '@src/http-exceptions/http-exception.module';
 import { InterceptorModule } from '@src/interceptors/interceptor.module';
@@ -20,6 +21,7 @@ import { UseDevelopmentMiddleware } from '@src/middlewares/use-development.middl
     InterceptorModule,
     HttpExceptionModule,
   ],
+  providers: [AppService],
 })
 export class AppModule implements NestModule {
   configure(consumer: MiddlewareConsumer) {
